python/requirements.txt
Django==1.6.6
gunicorn
dj-static
psycopg2
dj-database-url
South
python-social-auth
djangorestframework